Develop

+++ BusinessLogic/SubSystem/CategoriesSubSystem.cs
@@ -1,8 +1,8 @@
-﻿using BusinessLogic.Dominio;
-using BusinessLogic.Repository;
+﻿using BusinessLogic.Repository;
 using BusinessLogic.DTOs.DTOsCategory;
 using BusinessLogic.DTOs.DTOsSubCategory;
 using BusinessLogic.Común.Mappers;
+using BusinessLogic.Dominio;
 
 namespace BusinessLogic.SubSystem
 {
@@ -10,11 +10,13 @@ namespace BusinessLogic.SubSystem
     {
         private readonly ICategoryRepository _categoryRepository;
         private readonly ISubCategoryRepository _subCategoryRepository;
+        private readonly IProductRepository _productRepository;
 
-        public CategoriesSubSystem(ICategoryRepository categoryRepository, ISubCategoryRepository subCategoryRepository)
+        public CategoriesSubSystem(ICategoryRepository categoryRepository, ISubCategoryRepository subCategoryRepository, IProductRepository productRepository)
         {
             _categoryRepository = categoryRepository;
             _subCategoryRepository = subCategoryRepository;
+            _productRepository = productRepository;
         }
 
         // Categorías
@@ -27,7 +29,7 @@ namespace BusinessLogic.SubSystem
             var category = CategoryMapper.ToDomain(request);
             category.Validate();
 
-            var added = _categoryRepository.Add(category);
+            Category added = _categoryRepository.Add(category);
             return CategoryMapper.ToResponse(added);
         }
 
@@ -42,14 +44,19 @@ namespace BusinessLogic.SubSystem
             var updatedData = CategoryMapper.ToUpdatableData(request);
             existing.Update(updatedData);
 
-            var updated = _categoryRepository.Update(existing);
+            Category updated = _categoryRepository.Update(existing);
             return CategoryMapper.ToResponse(updated);
         }
 
         public CategoryResponse DeleteCategory(int id)
         {
-            var deleted = _categoryRepository.Delete(id)
+            Category deleted = _categoryRepository.GetById(id)
                           ?? throw new ArgumentException("Categoría no encontrada.", nameof(id));
+
+            if(deleted.SubCategories.Any())
+                throw new InvalidOperationException("No se puede eliminar una categoría que tiene subcategorías asociadas.");
+
+            _categoryRepository.Delete(id);
 
             return CategoryMapper.ToResponse(deleted);
         }
@@ -82,7 +89,7 @@ namespace BusinessLogic.SubSystem
             var subCategory = SubCategoryMapper.ToDomain(request);
             subCategory.Validate();
 
-            var added = _subCategoryRepository.Add(subCategory);
+            SubCategory added = _subCategoryRepository.Add(subCategory);
             return SubCategoryMapper.ToResponse(added);
         }
 
@@ -94,17 +101,22 @@ namespace BusinessLogic.SubSystem
             if (existing.Name != request.Name && _subCategoryRepository.GetByName(request.Name) != null)
                 throw new ArgumentException("Ya existe una subcategoría con el mismo nombre.", nameof(request.Name));
 
-            var updatedData = SubCategoryMapper.ToUpdatableData(request);
+            SubCategory.UpdatableData updatedData = SubCategoryMapper.ToUpdatableData(request);
             existing.Update(updatedData);
 
-            var updated = _subCategoryRepository.Update(existing);
+            SubCategory updated = _subCategoryRepository.Update(existing);
             return SubCategoryMapper.ToResponse(updated);
         }
 
         public SubCategoryResponse DeleteSubCategory(int id)
         {
-            var deleted = _subCategoryRepository.Delete(id)
+            var deleted = _subCategoryRepository.GetById(id)
                            ?? throw new ArgumentException("Subcategoría no encontrada.", nameof(id));
+
+            if (_productRepository.GetBySubCategoryId(id).Count() > 0)
+                throw new InvalidOperationException("No se puede eliminar una subcategoría que tiene productos asociados.");
+
+            _subCategoryRepository.Delete(id);
 
             return SubCategoryMapper.ToResponse(deleted);
         }
